python-agent-platform/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
sqlalchemy==2.0.38
pydantic==2.10.6
pydantic-settings==2.8.0
PyJWT==2.10.1
httpx==0.28.1
requests==2.32.3
Django==5.1.6
python-dotenv==1.0.1
pytest==8.3.5
ruff==0.9.7
mypy==1.15.0
types-requests==2.32.0.20250328
